backend/requirements.txt
# Core Framework
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
python-multipart==0.0.21

# MongoDB
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Validation & Settings
pydantic==2.12.5
pydantic-settings==2.2.1
pydantic_core==2.41.5
email-validator==2.3.0
annotated-types==0.7.0

# Authentication
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==46.0.3

# Data Processing
pandas==2.3.3
openpyxl==3.1.5
xlrd==2.0.2
PyPDF2==3.0.1
pdfplumber==0.11.0
tabula-py==2.9.0

# Environment & Utils
python-dotenv==1.2.1
typing_extensions==4.15.0

# Cloud Storage
b2sdk==2.5.0

# HTTP
anyio==4.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
sniffio==1.3.1
idna==3.11
certifi==2026.1.4
charset-normalizer==3.4.4
urllib3==2.6.2

# Async Support
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
frozenlist==1.8.0
multidict==6.7.0
yarl==1.22.0
attrs==25.4.0

# Other dependencies
click==8.3.1
numpy==2.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
six==1.17.0
cffi==2.0.0
pycparser==2.23
